add NextOperation
